feat: add computeDefaultLayout and associated utility functions for layout management
