+++ packages/layout/src/node-renderers/split-node-renderer/SplitLayoutPanelGroup.tsx
@@ -9,7 +9,7 @@ import {useSplitNodeContext} from '../../LayoutNodeContext';
 import {useLayoutRendererContext} from '../../LayoutRendererContext';
 import {updateLayoutNodeById} from '../../layout-tree';
 import type {Layout, LayoutChangedMeta} from 'react-resizable-panels';
-import {isCollapsed} from '../utils';
+import {computeDefaultLayout} from './computeDefaultLayout';
 
 export type RenderPanelProps = {
   node: LayoutNode;
@@ -54,117 +54,6 @@ function getPanelSizeInPixels(
   return direction === 'column'
     ? panelElement?.offsetHeight
     : panelElement?.offsetWidth;
-}
-
-/**
- * Parse a node `defaultSize` into a percentage number, or `undefined` when it
- * is not expressed as a percentage (e.g. pixels or unset). Pixel/unset sizes
- * cannot be converted without measuring the container, so they are treated as
- * "unknown" and distributed evenly by {@link computeDefaultLayout}.
- */
-function parsePercentSize(
-  size: string | number | undefined,
-): number | undefined {
-  if (typeof size === 'number') {
-    return undefined; // numeric == pixels in this codebase
-  }
-  if (typeof size === 'string' && size.trim().endsWith('%')) {
-    const value = Number.parseFloat(size);
-    return Number.isFinite(value) ? value : undefined;
-  }
-  return undefined;
-}
-
-/**
- * Whether a node `defaultSize` carries an explicit pixel intent (a raw number
- * or a `px` string). Such sizes cannot be represented in the group-level,
- * percentage-only RRP `Layout`, so a group containing one must not emit a
- * `defaultLayout` (see {@link computeDefaultLayout}).
- */
-function isPixelSize(size: string | number | undefined): boolean {
-  if (typeof size === 'number') {
-    return true; // numeric == pixels in this codebase
-  }
-  return typeof size === 'string' && size.trim().endsWith('px');
-}
-
-/**
- * Derive the group's initial RRP `Layout` ({panelId → percentage}) from the
- * declared node sizes and collapsed flags. Collapsed nodes get 0; the rest fill
- * the remaining space proportionally to their `defaultSize` (unset sizes share
- * evenly). This makes RRP paint the correct (already-collapsed) layout on the
- * very first frame instead of flashing the uncollapsed sizes and then
- * imperatively collapsing after mount.
- *
- * Returns `undefined` (so RRP honours each panel's own `defaultSize`) when a
- * VISIBLE panel declares a pixel size: the group `Layout` is percentage-only,
- * so emitting it would silently rewrite e.g. a 250px sidebar to a percentage.
- * A pixel panel that is currently collapsed is not "visible", so groups still
- * get their anti-flash layout while such a panel stays collapsed.
- */
-export function computeDefaultLayout(
-  children: LayoutNode[],
-): Layout | undefined {
-  if (children.length === 0) {
-    return undefined;
-  }
-
-  const entries = children.map((child) => {
-    const id = getLayoutNodeId(child);
-    const collapsed = isCollapsed(child);
-    const isKey = isLayoutNodeKey(child);
-    const percent = isKey ? undefined : parsePercentSize(child.defaultSize);
-    const pixel = isKey ? false : isPixelSize(child.defaultSize);
-    return {id, collapsed, percent, pixel};
-  });
-
-  const visible = entries.filter((entry) => !entry.collapsed);
-  if (visible.length === 0) {
-    // Degenerate: everything collapsed. Let RRP fall back to per-panel sizing.
-    return undefined;
-  }
-
-  // A visible pixel-sized panel cannot be faithfully encoded as a percentage.
-  // Defer entirely to per-panel `defaultSize` rather than distort it.
-  if (visible.some((entry) => entry.pixel)) {
-    return undefined;
-  }
-
-  const knownSum = visible.reduce(
-    (sum, entry) => sum + (entry.percent ?? 0),
-    0,
-  );
-  const unknownCount = visible.filter(
-    (entry) => entry.percent === undefined,
-  ).length;
-  const remaining = Math.max(0, 100 - knownSum);
-  const perUnknown = unknownCount > 0 ? remaining / unknownCount : 0;
-
-  const rawWeights = new Map<string, number>();
-  for (const entry of entries) {
-    if (entry.collapsed) {
-      rawWeights.set(entry.id, 0);
-    } else {
-      rawWeights.set(entry.id, entry.percent ?? perUnknown);
-    }
-  }
-
-  const visibleTotal = visible.reduce(
-    (sum, entry) => sum + (rawWeights.get(entry.id) ?? 0),
-    0,
-  );
-
-  const layout: Layout = {};
-  for (const entry of entries) {
-    if (entry.collapsed) {
-      layout[entry.id] = 0;
-    } else if (visibleTotal > 0) {
-      layout[entry.id] = ((rawWeights.get(entry.id) ?? 0) / visibleTotal) * 100;
-    } else {
-      layout[entry.id] = 100 / visible.length;
-    }
-  }
-  return layout;
 }
 
 function updateNodeDefaultSize(
